[NF] OptiObjective class
