+++ Tests/Plot/test_ICEM_2020.py
@@ -20,7 +20,7 @@ from pyleecan.Classes.MagFEMM import MagFEMM
 from pyleecan.Classes.Simu1 import Simu1
 from pyleecan.Classes.Output import Output
 from pyleecan.Classes.OptiDesignVar import OptiDesignVar
-from pyleecan.Classes.DataKeeper import DataKeeper
+from pyleecan.Classes.OptiObjective import OptiObjective
 from pyleecan.Classes.OptiProblem import OptiProblem
 from pyleecan.Classes.ImportMatrixVal import ImportMatrixVal
 from pyleecan.Classes.ImportGenVectLin import ImportGenVectLin
@@ -720,13 +720,13 @@ def test_Optimization_problem():
     Tem_rip_pp = "lambda output: abs(output.mag.Tem_rip_pp)"
 
     my_objs = [
-        DataKeeper(
+        OptiObjective(
             name="Maximization of the average torque",
             symbol="Tem_av",
             unit="N.m",
             keeper=tem_av,
         ),
-        DataKeeper(
+        OptiObjective(
             name="Minimization of the torque ripple",
             symbol="Tem_rip_pp",
             unit="N.m",
