+++ Tests/Validation/Optimization/test_opti_preprocessing.py
@@ -12,7 +12,7 @@ import pytest
 from pyleecan.Classes.Simu1 import Simu1
 from pyleecan.Classes.Output import Output
 from pyleecan.Classes.OptiDesignVar import OptiDesignVar
-from pyleecan.Classes.DataKeeper import DataKeeper
+from pyleecan.Classes.OptiObjective import DatOptiObjectiveaKeeper
 from pyleecan.Classes.OptiProblem import OptiProblem
 from pyleecan.Classes.ImportMatrixVal import ImportMatrixVal
 from pyleecan.Classes.ImportGenVectLin import ImportGenVectLin
@@ -57,7 +57,7 @@ def test_opti_preprocessing():
 
     # Objectives
     objs = [
-        DataKeeper(
+        OptiObjective(
             name="Minimization of the rotor slot width",
             symbol="R_s_w0",
             unit="m",
