+++ pyleecan/Methods/Output/XOutput/plot_multi.py
@@ -2,6 +2,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from ....definitions import config_dict
+from ....Methods.Output.XOutput import _get_symbol_data_
 
 COLORS = config_dict["PLOT"]["COLOR_DICT"]["CURVE_COLORS"]
 FONT_NAME = config_dict["PLOT"]["FONT_NAME"]
@@ -45,31 +46,9 @@ def plot_multi(
     if plot_type not in ["scatter", "plot"]:
         raise PlotError("Unknown plot_type {}.".format(plot_type))
 
-    # Get x_data
-    if x_symbol in self.keys():  # DataKeeper
-        x_data = self[x_symbol]
-        x_values = np.array(x_data.result)[idx]
-    else:  # ParamSetter
-        x_data = self.get_paramexplorer(x_symbol)
-        x_values = np.array(x_data.value)[idx]
-
-    # x_label definition
-    x_label = x_symbol
-    if x_data.unit not in ["", None]:
-        x_label += " [{}]".format(x_data.unit)
-
-    # Get y_data
-    if y_symbol in self.keys():  # DataKeeper
-        y_data = self[y_symbol]
-        y_values = np.array(y_data.result)[idx]
-    else:  # ParamSetter
-        y_data = self.get_paramexplorer(y_symbol)
-        y_values = np.array(y_data.value)[idx]
-
-    # y_label definition
-    y_label = y_symbol
-    if y_data.unit not in ["", None]:
-        y_label += " [{}]".format(y_data.unit)
+    # get data and labels
+    x_values, x_label = _get_symbol_data_(self, x_symbol, idx)
+    y_values, y_label = _get_symbol_data_(self, y_symbol, idx)
 
     # Plot in new figure
     if ax is None:
